Re-implement ProfileDataSource as an interface
